refactor: replace serial, VGA, and panic handling with petroleum crate
- Remove custom serial port implementation in fullerene-kernel/src/serial.rs and delegate to petroleum::serial
- Remove custom VGA text mode structures and writer in fullerene-kernel/src/vga.rs, using petroleum::vga instead
- Update panic handler in bellows/src/main.rs to use petroleum::handle_panic instead of print-and-loop logic
- Reduces code duplication and leverages shared petroleum library for EFI and kernel I/O operations
