+++ bellows/src/main.rs
@@ -143,12 +143,7 @@ pub extern "efiapi" fn efi_main(image_handle: usize, system_table: *mut EfiSyste
     }
 }
 
-#[cfg(not(test))]
-#[panic_handler]
-fn panic(info: &core::panic::PanicInfo) -> ! {
-    petroleum::println!("PANIC: {}", info);
-    loop {}
-}
+
 
 /// Initializes the Graphics Output Protocol (GOP) for framebuffer access.
 fn init_gop(st: &EfiSystemTable) {
@@ -223,8 +218,8 @@ fn init_gop(st: &EfiSystemTable) {
     }
 }
 
-// #[cfg(not(test))]
-// #[panic_handler]
-// fn panic(info: &core::panic::PanicInfo) -> ! {
-//     petroleum::handle_panic(info)
-// }
+#[cfg(not(test))]
+#[panic_handler]
+fn panic(info: &core::panic::PanicInfo) -> ! {
+    petroleum::handle_panic(info)
+}
