+++ petroleum/src/lib.rs
@@ -5,8 +5,11 @@
 extern crate alloc;
 
 pub mod common;
+pub mod graphics;
 pub mod page_table;
 pub mod serial;
+pub use graphics::{Color, ColorCode, ScreenChar, TextBufferOperations};
+pub use serial::{SerialPortWriter, Com1Ops, SerialPort, SerialPortOps};
 
 use core::{arch::asm, fmt::Write};
 use spin::Mutex;
